refactored combining projects into its own file
